Cambios en DAOs y se agrega pantalla ConsultaSucursal

+++ died-tp/src/main/java/datos/Ruta.java
@@ -6,11 +6,11 @@ public class Ruta {
 	private Sucursal destino;
 	private Operatividad estado;
 	private Integer duracion; //Unidad: minutos
-	private Double capacidadMaxima; //Unidad: kg
+	private Float capacidadMaxima; //Unidad: kg
 	
 	public Sucursal getOrigen() {return this.origen;}
 	public Sucursal getDestino() {return this.destino;}
 	public Integer getDuracion() {return this.duracion;}
 	public String getIdRuta() {return this.idRuta;}
-	public Double getCapacidadMaxima() {return this.capacidadMaxima;}
+	public Float getCapacidadMaxima() {return this.capacidadMaxima;}
 }
